+++ app/src/main/java/it/unibo/unibodget/model/dashboard/api/BudgetMonitor.java
@@ -1,5 +1,7 @@
 package it.unibo.unibodget.model.dashboard.api;
 
+import java.math.BigDecimal;
+
 /**
  * Service responsible for evaluating the current budget condition.
  * This component compares the current aggregated dashboard value with the
@@ -13,5 +15,6 @@ public interface BudgetMonitor {
      * @param settings the user-defined budget settings used during the evaluation
      * @return the resulting budget status
      */
-    BudgetStatus getBudgetStatus(double currentValue, BudgetSettings settings);
+    BudgetStatus getBudgetStatus(BigDecimal currentValue, BudgetSettings settings);
 }
+
