+++ app/src/main/java/it/unibo/unibodget/model/dashboard/impl/CriticalBudgetStrategy.java
@@ -1,5 +1,7 @@
 package it.unibo.unibodget.model.dashboard.impl;
 
+import java.math.BigDecimal;
+
 import it.unibo.unibodget.model.dashboard.api.BudgetAlertStrategy;
 import it.unibo.unibodget.model.dashboard.api.BudgetStatus;
 
@@ -12,7 +14,7 @@ public final class CriticalBudgetStrategy implements BudgetAlertStrategy {
      * {@inheritDoc}
      */
     @Override
-    public BudgetStatus evaluate(final double currentValue, final double limitValue) {
+    public BudgetStatus evaluate(final BigDecimal currentValue, final BigDecimal limitValue) {
         return BudgetStatus.CRITICAL;
     }
 }
