+++ app/src/main/java/it/unibo/unibodget/model/dashboard/impl/DefaultDashboardSnapshot.java
@@ -1,5 +1,6 @@
 package it.unibo.unibodget.model.dashboard.impl;
 
+import java.math.BigDecimal;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
@@ -13,9 +14,9 @@ import it.unibo.unibodget.model.transactions.base.Transaction;
  */
 public final class DefaultDashboardSnapshot implements DashboardSnapshot {
 
-    private final double totalBalance;
+    private final BigDecimal totalBalance;
     private final List<Transaction> recentTransactions;
-    private final Map<String, Double> categorySummaries;
+    private final Map<String, BigDecimal> categorySummaries;
     private final BudgetStatus budgetStatus;
 
     /**
@@ -27,9 +28,9 @@ public final class DefaultDashboardSnapshot implements DashboardSnapshot {
      * @param budgetStatus the current budget status
      */
     public DefaultDashboardSnapshot(
-            final double totalBalance,
+            final BigDecimal totalBalance,
             final List<Transaction> recentTransactions,
-            final Map<String, Double> categorySummaries,
+            final Map<String, BigDecimal> categorySummaries,
             final BudgetStatus budgetStatus) {
         this.totalBalance = totalBalance;
         this.recentTransactions = List.copyOf(recentTransactions);
@@ -41,7 +42,7 @@ public final class DefaultDashboardSnapshot implements DashboardSnapshot {
      * {@inheritDoc}
      */
     @Override
-    public double getTotalBalance() {
+    public BigDecimal getTotalBalance() {
         return totalBalance;
     }
 
@@ -57,7 +58,7 @@ public final class DefaultDashboardSnapshot implements DashboardSnapshot {
      * {@inheritDoc}
      */
     @Override
-    public Map<String, Double> getCategorySummaries() {
+    public Map<String, BigDecimal> getCategorySummaries() {
         return Collections.unmodifiableMap(categorySummaries);
     }
